+++ src/components/FacetOptionFormatter/index.js
@@ -1,0 +1,2 @@
+export { default } from './FacetOptionFormatter';
+export { default as FacetOptionFormatter } from './FacetOptionFormatter';