+++ src/components/MultiSelectionFacet/index.js
@@ -1,0 +1,2 @@
+export { default } from './MultiSelectionFacet';
+export { default as MultiSelectionFacet } from './MultiSelectionFacet';