+++ CHANGELOG.md
@@ -53,7 +53,10 @@ Inventory causes an error. Refs UIIN-2012.
 * Browse contributors list. Refs UIIN-2024.
 * Normalize ISBNs for ISBN searching. Search option: Keyword (Instance, Holdings, Item). Refs UIIN-998.
 * Normalize ISBNs for ISBN searching. Search option: Identifier (all) in the Instance segment. Refs UIIN-999.
-
+* Browse form - first and last record navigation. Refs UIIN-1912.
+* Inventory >  Contributors Browse - Create a Name Type facet. Refs UIIN-2021.
+* Add filter/facet for Holdings > holdings type. Refs UIIN-1347.
+* Fix when click view Holdings the Something went wrong error page appears. Refs UIIN-2063.
 
 ## [9.0.0](https://github.com/folio-org/ui-inventory/tree/v9.0.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v8.0.0...v9.0.0)